requirements.txt
# Core dependencies for V3 Pipeline
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
torch>=1.9.0
transformers>=4.20.0

# LLM Provider APIs
openai>=1.0.0
anthropic>=0.3.0
google-genai>=0.3.0

# Consensus and clustering algorithms
hdbscan>=0.8.28
editdistance>=0.6.0
networkx>=2.6.0

# Utilities and data processing
pathlib2>=2.3.6
tqdm>=4.62.0
requests>=2.26.0
aiohttp>=3.8.0

# Optional: For advanced features
asyncio-mqtt>=0.10.0
matplotlib>=3.5.0
seaborn>=0.11.0